reorganize frontend components
