+++ frontend/src/components/chat/index.js
@@ -1,0 +1,2 @@
+import Chat from "./Chat";
+export default Chat;
