+++ frontend/src/components/chat/styles/ChatStyles.js
