+++ frontend/src/components/game/index.js
@@ -1,0 +1,2 @@
+import GameContainer from "./GameContainer";
+export default GameContainer;
