+++ frontend/src/components/homepage/index.js
@@ -1,0 +1,2 @@
+import Welcome from "./Welcome";
+export default Welcome;
