+++ frontend/src/components/shared/ErrorBoundary.js
