+++ frontend/src/components/shared/__tests__/AnimatedNumber.test.js
