+++ frontend/src/components/shared/__tests__/Error.test.js
