+++ frontend/src/components/shared/wordgrid/TouchListener.js
