+++ frontend/src/components/shared/wordgrid/index.js
@@ -1,0 +1,2 @@
+import WordGrid from "./WordGrid";
+export default WordGrid;
