+++ frontend/src/components/shared/wordgrid/utils.js
